--- a/aten/src/ATen/native/cuda/thread_constants.h
+++ b/aten/src/ATen/native/cuda/thread_constants.h
@@ -18,5 +18,5 @@ constexpr uint32_t num_threads() {
 }
 #endif
 
-constexpr int thread_work_size() { return 4; }
+constexpr int thread_work_size() { return 8; }
 constexpr int block_work_size() { return thread_work_size() * num_threads(); }