@ngimel 2024-12-03
This unconditionally sets items per thread to 8. Is it neutral/beneficial on A100?