@Aidyn-A 2024-12-05
> This unconditionally sets items per thread to 8. Is it neutral/beneficial on A100?

It improves performance on A100 as well.

<details><summary> A100 benchmark results are here </summary>

 | op | dtype | size | time after | time before | % improvement |
 | ---- | ---- | ---- | ---- | ---- | ---- |
 | relu | torch.float16 | 16777216 | 5.52E-05 | 5.56E-05 | 0.588398905051735 | 
 | relu | torch.float16 | 33554432 | 0.000101789004272885 | 0.000103080593463447 | 1.26888871719382 | 
 | relu | torch.float16 | 67108864 | 0.000197273265156481 | 0.000199393058816592 | 1.07454685176394 | 
 | relu | torch.float16 | 134217728 | 0.000388443004339933 | 0.000392856593761179 | 1.13622574533054 | 
 | relu | torch.float16 | 268435456 | 0.00077106234514051 | 0.000779860394282474 | 1.14102954156841 | 
 | relu | torch.float16 | 536870912 | 0.00153737427252862 | 0.00155346540527211 | 1.04666332922463 | 
 | relu | torch.bfloat16 | 16777216 | 5.50E-05 | 5.44E-05 | -1.00206226691304 | 
 | relu | torch.bfloat16 | 33554432 | 0.000104047668476899 | 0.000104331825342443 | 0.273102578561746 | 
 | relu | torch.bfloat16 | 67108864 | 0.000202173677583536 | 0.000203764117840264 | 0.786670290483738 | 
 | relu | torch.bfloat16 | 134217728 | 0.000398641969594691 | 0.000402679149475363 | 1.01273327662332 | 
 | relu | torch.bfloat16 | 268435456 | 0.000800862442702055 | 0.000801527769201332 | 0.0830762517757711 | 
 | relu | torch.bfloat16 | 536870912 | 0.00159262172463867 | 0.00157401508962115 | -1.16830222328839 | 
 | relu | torch.float32 | 16777216 | 0.000105040096160438 | 0.000104826926771137 | -0.202940969299259 | 
 | relu | torch.float32 | 33554432 | 0.000205832430058055 | 0.000205058294037978 | -0.376100121763734 | 
 | relu | torch.float32 | 67108864 | 0.000406690562764804 | 0.000405893247160647 | -0.196049694081057 | 
 | relu | torch.float32 | 134217728 | 0.000808526658349567 | 0.000807985611673858 | -0.0669176050191012 | 
 | relu | torch.float32 | 268435456 | 0.00159855828517013 | 0.00159612013440993 | -0.152521855650756 | 
 | relu | torch.float32 | 536870912 | 0.00310465165724357 | 0.00309997693532043 | -0.150571543581468 | 
 | sigmoid | torch.float16 | 16777216 | 5.53E-05 | 5.60E-05 | 1.37442506088692 | 
 | sigmoid | torch.float16 | 33554432 | 0.000104938064598375 | 0.000106826372858551 | 1.79945024467809 | 
 | sigmoid | torch.float16 | 67108864 | 0.000204297041313516 | 0.000208476403107246 | 2.04572800803136 | 
 | sigmoid | torch.float16 | 134217728 | 0.0004032871996363 | 0.000411878443426556 | 2.13030411032227 | 
 | sigmoid | torch.float16 | 268435456 | 0.000802099549522003 | 0.000818967301812437 | 2.10294997678111 | 
 | sigmoid | torch.float16 | 536870912 | 0.00159777545680602 | 0.00163260997376508 | 2.18018851213879 | 
 | sigmoid | torch.bfloat16 | 16777216 | 5.53E-05 | 5.60E-05 | 1.27645767333306 | 
 | sigmoid | torch.bfloat16 | 33554432 | 0.00010527613469296 | 0.000107017759647634 | 1.65433976062424 | 
 | sigmoid | torch.bfloat16 | 67108864 | 0.000203829206940201 | 0.000208569638844993 | 2.3256882445623 | 
 | sigmoid | torch.bfloat16 | 134217728 | 0.000402993574324581 | 0.00041222023881144 | 2.28953141556227 | 
 | sigmoid | torch.bfloat16 | 268435456 | 0.0008021114497549 | 0.000818662500629822 | 2.06343530939241 | 
 | sigmoid | torch.bfloat16 | 536870912 | 0.00159755944170886 | 0.00163286800185839 | 2.21015626885024 | 
 | sigmoid | torch.float32 | 16777216 | 0.000105733621037669 | 0.000104870129790571 | -0.816666674822408 | 
 | sigmoid | torch.float32 | 33554432 | 0.000206765667017963 | 0.000204619951546192 | -1.03775230323157 | 
 | sigmoid | torch.float32 | 67108864 | 0.000407101896901925 | 0.000405989535566833 | -0.273239044955886 | 
 | sigmoid | torch.float32 | 134217728 | 0.000807929680579238 | 0.000807615410950449 | -0.0388981413041223 | 
 | sigmoid | torch.float32 | 268435456 | 0.00161020312872198 | 0.00161120223088397 | 0.0620482064758621 | 
 | sigmoid | torch.float32 | 536870912 | 0.00312601159223252 | 0.00311755073360271 | -0.270659860981737 | 
 | tanh | torch.float16 | 16777216 | 5.50E-05 | 5.56E-05 | 1.01392536515987 | 
 | tanh | torch.float16 | 33554432 | 0.00010461484392484 | 0.000106048201107317 | 1.37012791751241 | 
 | tanh | torch.float16 | 67108864 | 0.000203798576775524 | 0.000206987477011151 | 1.56473135685284 | 
 | tanh | torch.float16 | 134217728 | 0.000403522617287106 | 0.000408950727432966 | 1.3451811406145 | 
 | tanh | torch.float16 | 268435456 | 0.000802456504768795 | 0.000812884057975478 | 1.29945400712861 | 
 | tanh | torch.float16 | 536870912 | 0.0015979148964915 | 0.00162050236637394 | 1.41355900317532 | 
 | tanh | torch.bfloat16 | 16777216 | 5.50E-05 | 5.58E-05 | 1.32688536364782 | 
 | tanh | torch.bfloat16 | 33554432 | 0.000104613498681121 | 0.000106201765851842 | 1.51822392974537 | 
 | tanh | torch.bfloat16 | 67108864 | 0.00020353009717332 | 0.000207337292118205 | 1.87058081225351 | 
 | tanh | torch.bfloat16 | 134217728 | 0.000403028550661272 | 0.000410060760461622 | 1.74484159715513 | 
 | tanh | torch.bfloat16 | 268435456 | 0.00080103049468663 | 0.000814526652296384 | 1.68484941575588 | 
 | tanh | torch.bfloat16 | 536870912 | 0.00159850344061852 | 0.0016241325582895 | 1.6033195187287 | 
 | tanh | torch.float32 | 16777216 | 0.000105516829838355 | 0.000104838620043463 | -0.642750351703303 | 
 | tanh | torch.float32 | 33554432 | 0.000205859490152862 | 0.000204412835753626 | -0.702738745812404 | 
 | tanh | torch.float32 | 67108864 | 0.000406586409856876 | 0.00040571458844675 | -0.214424631269161 | 
 | tanh | torch.float32 | 134217728 | 0.000808000409354766 | 0.000807130088408788 | -0.107712933793336 | 
 | tanh | torch.float32 | 268435456 | 0.00160988306419717 | 0.00161015692477425 | 0.017011209271689 | 
 | tanh | torch.float32 | 536870912 | 0.00313035196935137 | 0.00310210293779771 | -0.90242349199855 | 
 | gelu | torch.float16 | 16777216 | 5.58E-05 | 5.70E-05 | 2.15048278510104 | 
 | gelu | torch.float16 | 33554432 | 0.000105667704095443 | 0.000109181273728609 | 3.32511211750415 | 
 | gelu | torch.float16 | 67108864 | 0.000205330136749479 | 0.000213289943834146 | 3.87658977424148 | 
 | gelu | torch.float16 | 134217728 | 0.000404131753991048 | 0.000421578219781319 | 4.31702424221241 | 
 | gelu | torch.float16 | 268435456 | 0.00080191721725795 | 0.000838274239665932 | 4.53376254126332 | 
 | gelu | torch.float16 | 536870912 | 0.00159792177793052 | 0.00167205422702763 | 4.63930400855495 | 
 | gelu | torch.bfloat16 | 16777216 | 5.56E-05 | 5.71E-05 | 2.7460248048456 | 
 | gelu | torch.bfloat16 | 33554432 | 0.000105594181352192 | 0.000109837338742283 | 4.01836288302555 | 
 | gelu | torch.bfloat16 | 67108864 | 0.000204587148295508 | 0.000214392060620917 | 4.79253580056083 | 
 | gelu | torch.bfloat16 | 134217728 | 0.000403126029090749 | 0.000423436725719108 | 5.03829948023189 | 
 | gelu | torch.bfloat16 | 268435456 | 0.000799921082539691 | 0.000842164632760816 | 5.28096472804605 | 
 | gelu | torch.bfloat16 | 536870912 | 0.00159319071099162 | 0.00167886317811078 | 5.37741442553605 | 
 | gelu | torch.float32 | 16777216 | 0.00010601943358779 | 0.000104866456240416 | -1.08751509827605 | 
 | gelu | torch.float32 | 33554432 | 0.000206458692749341 | 0.000204429340859254 | -0.982933614013981 | 
 | gelu | torch.float32 | 67108864 | 0.000406732886201806 | 0.000404650604145394 | -0.511953207387805 | 
 | gelu | torch.float32 | 134217728 | 0.00080782449286845 | 0.000810090193731917 | 0.280469443978126 | 
 | gelu | torch.float32 | 268435456 | 0.00160836293879482 | 0.00160547635621495 | -0.179473333428126 | 
 | gelu | torch.float32 | 536870912 | 0.00312541264833676 | 0.00309180561453104 | -1.07528309337327 | 
 | sin | torch.float16 | 16777216 | 5.58E-05 | 5.68E-05 | 1.84432136815518 | 
 | sin | torch.float16 | 33554432 | 0.000105774857931667 | 0.000109315746360355 | 3.34757096149971 | 
 | sin | torch.float16 | 67108864 | 0.000204949174076319 | 0.000213460582825873 | 4.15293635015306 | 
 | sin | torch.float16 | 134217728 | 0.000403335266229179 | 0.000421439038796557 | 4.48851714273144 | 
 | sin | torch.float16 | 268435456 | 0.000803771014842722 | 0.00083896823020445 | 4.37901027926653 | 
 | sin | torch.float16 | 536870912 | 0.00159420336907109 | 0.00167318914706508 | 4.95456097549341 | 
 | sin | torch.bfloat16 | 16777216 | 5.58E-05 | 5.70E-05 | 2.05289950280472 | 
 | sin | torch.bfloat16 | 33554432 | 0.000105607685529523 | 0.000109381404601865 | 3.57333754018021 | 
 | sin | torch.bfloat16 | 67108864 | 0.000205246110757192 | 0.000213417379806439 | 3.98120530474448 | 
 | sin | torch.bfloat16 | 134217728 | 0.000403606125877963 | 0.000421721229536666 | 4.48831236624492 | 
 | sin | torch.bfloat16 | 268435456 | 0.000800546879569689 | 0.000839088215596146 | 4.81437589853244 | 
 | sin | torch.bfloat16 | 536870912 | 0.00159501155010528 | 0.0016719801351428 | 4.8255816725868 | 
 | sin | torch.float32 | 16777216 | 0.0001063231482274 | 0.000104805506351921 | -1.42738613442208 | 
 | sin | torch.float32 | 33554432 | 0.000206721118754811 | 0.00020440181510316 | -1.1219480939447 | 
 | sin | torch.float32 | 67108864 | 0.000406984757218096 | 0.000404441884408395 | -0.624807874153088 | 
 | sin | torch.float32 | 134217728 | 0.000807934802853399 | 0.000804758371992244 | -0.393154354774283 | 
 | sin | torch.float32 | 268435456 | 0.00160756231182151 | 0.00160335816649927 | -0.261523008553233 | 
 | sin | torch.float32 | 536870912 | 0.00312547411562668 | 0.00309828793009122 | -0.869825969747518 | 
 | exp | torch.float16 | 16777216 | 5.46E-05 | 5.49E-05 | 0.526774300459265 | 
 | exp | torch.float16 | 33554432 | 0.000104664100541009 | 0.000104676155994336 | 0.0115182314324169 | 
 | exp | torch.float16 | 67108864 | 0.000204241420659754 | 0.000204416768004497 | 0.0858529793694496 | 
 | exp | torch.float16 | 134217728 | 0.000404592706925339 | 0.000404040846559736 | -0.136398989936448 | 
 | exp | torch.float16 | 268435456 | 0.000805885789708959 | 0.00080300003497137 | -0.35808482720997 | 
 | exp | torch.float16 | 536870912 | 0.00160873257037666 | 0.00160096456400222 | -0.482864990581999 | 
 | exp | torch.bfloat16 | 16777216 | 5.49E-05 | 5.47E-05 | -0.489032335861872 | 
 | exp | torch.bfloat16 | 33554432 | 0.000104703112608857 | 0.0001046407657365 | -0.0595463408897046 | 
 | exp | torch.bfloat16 | 67108864 | 0.000204433790511555 | 0.000204202770772907 | -0.113004674065675 | 
 | exp | torch.bfloat16 | 134217728 | 0.000404823674923844 | 0.000403735734936264 | -0.268744160722478 | 
 | exp | torch.bfloat16 | 268435456 | 0.000805160548124048 | 0.000802269826332728 | -0.359024271377362 | 
 | exp | torch.bfloat16 | 536870912 | 0.00160597870126367 | 0.00160156599142485 | -0.274767643888552 | 
 | exp | torch.float32 | 16777216 | 0.000105293364160591 | 0.00010481502653824 | -0.454290378282118 | 
 | exp | torch.float32 | 33554432 | 0.000205790779242913 | 0.000204942189157009 | -0.412355737718517 | 
 | exp | torch.float32 | 67108864 | 0.000406338781532314 | 0.000405922169900603 | -0.102528149082948 | 
 | exp | torch.float32 | 134217728 | 0.000808410087807311 | 0.000807947479188442 | -0.057224498536812 | 
 | exp | torch.float32 | 268435456 | 0.00161091559049156 | 0.00161218053350846 | 0.0785232338904818 | 
 | exp | torch.float32 | 536870912 | 0.0031157395698958 | 0.00311509809560246 | -0.020588187136561 | 

</details>